update to latest

+++ c/modp_qsiter.c
@@ -73,14 +73,14 @@ bool qsiter_next(struct qsiter_t* qsi)
         char* eq = (char*) memchr(charstart, '=', qsi->len - qsi->pos);
         if (eq == NULL) {
             qsi->key = charstart;
-            qsi->keylen = qsi->len - qsi->pos;
+            qsi->keylen = (size_t)(qsi->len - qsi->pos);
             qsi->val = NULL;
-            qsi->vallen = 0;
+            qsi->vallen = (size_t)0;
         } else {
             qsi->key = charstart;
-            qsi->keylen = eq - charstart;
+            qsi->keylen = (size_t)(eq - charstart);
             qsi->val = eq + 1;
-            qsi->vallen = (qsi->s + qsi->len) - qsi->val;
+            qsi->vallen = (size_t)((qsi->s + qsi->len) - qsi->val);
         }
         qsi->pos = qsi->len;
         return true;
@@ -89,16 +89,16 @@ bool qsiter_next(struct qsiter_t* qsi)
         char* eq = (char*) memchr(charstart, '=', (size_t)(ends - charstart));
         if (eq == NULL) {
             qsi->key = charstart;
-            qsi->keylen = ends - charstart;
+            qsi->keylen = (size_t)(ends - charstart);
             qsi->val = NULL;
-            qsi->vallen = 0;
+            qsi->vallen = (size_t)0;
         } else {
             qsi->key = charstart;
-            qsi->keylen = eq - charstart;
+            qsi->keylen = (size_t)(eq - charstart);
             qsi->val = eq + 1;
-            qsi->vallen = ends - eq - 1;
+            qsi->vallen = (size_t)(ends - eq - 1);
         }
-        qsi->pos = (ends - qsi->s) + 1;
+        qsi->pos = (size_t)((ends - qsi->s) + 1);
         return true;
     }
 }
